requirements.txt
# Core Backend Dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Database
sqlalchemy>=2.0.0
pydantic>=2.5.0

# Frontend
streamlit>=1.29.0
streamlit-webrtc>=0.47.0

# Face Recognition & Computer Vision
face-recognition>=1.3.0
opencv-python>=4.8.0
opencv-contrib-python>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0
dlib>=19.24.0

# Data Processing
pandas>=2.1.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
httpx>=0.25.0

# Utilities
python-dotenv>=1.0.0
